Move the pip install's to their own file.
Summary:
This will make it easier to update the pip requirements once the
setup.sh functionality moved to the aws-repo instead of being in this
one.

Get the pip syntax right.

Test Plan:
None

pip install -r requirements.txt
and saw it start to run successfully (downloading stuff and the like)

Reviewers: tom

Reviewed By: tom

Subscribers: jlfwong

Differential Revision: http://phabricator.khanacademy.org/D14527