Low-level storage for errors from BigQuery logs.
Summary:
Before I dive into reimplementing the BigQuery importer logic, I figured I'd
start by implementing the persistence logic to get some feedback on how the
data will be stored in Redis and how it will interact with the existing
monitoring data. The two are currently separated by prepending "MON_" to the
version strings while monitoring.

I've also added an API endpoint to fetch summary error information so I can
start to unit test the save/load codepaths properly. This is probably all the
data that the UI will use to render the error summary, except that A) we'll
want to return all the errors for a version at once and B) we'll need to cache
the data for performance.

What needs to happen before this ships:
1. (Re)Implement BigQuery importer process & create a cron job
2. Use the BigQuery data from the known good deployed versions to avoid the
   log spam issues we were seeing when error-monitor-db was launched
3. Alert when new errors show up in the logs that don't show up for the first
   time during monitoring

Test Plan: Ran `python test_server.py`

Reviewers: jlfwong, csilvers

Reviewed By: jlfwong, csilvers

Subscribers: jace

Differential Revision: https://phabricator.khanacademy.org/D14885