Error monitor: Clean up how BigQuery import errors are handled.
Summary:
Previously, errors would be written to stdout but no error code would be
returned, so they won't be reported by cron. Now instead of returning None
values from `run_query` we raise an exception with the error information and
report it in __main__ as well as returning an error code.

Also, reaching a dataset that isn't in BigQuery yet no longer prints out a
scary-looking message since it happens most of the time. It also won't return
an error code.

Test Plan:
- Tweak the SQL to include a syntax error.
- Run `bigquery_import.py`
- See the error message reported nicely on stdout
- Fix the SQL
- Run `bigquery_import.py` again
- It should run successfully and report a non-scary-looking error message
  when it reaches a dataset that BigQuery doesn't have yet

Reviewers: jlfwong, csilvers

Reviewed By: csilvers

Subscribers: jace

Differential Revision: https://phabricator.khanacademy.org/D15105