Re-add error information to Redis if it has expired.
Summary:
If an error that has last occurred over a week ago (and thus expired from
Redis) recurs, we will find its signature in the 'errordef' hashtables (because
those don't expire) and previously we would not then add the error data to
Redis. This led to a crash in the bigquery_import script because it assumed
that all "new" errors were in fact saved in Redis. There are probably other
places where that assumption is made.

This change will fix things so that the error will be rewritten to Redis if
it has expired.

Test Plan: Ran `test_server.py`

Reviewers: jlfwong, csilvers

Reviewed By: csilvers

Subscribers: jace

Differential Revision: https://phabricator.khanacademy.org/D15774