Sort versions so that Dec. 2014 comes before Jan. 2015.
Summary:
This is a hack because we don't have year information in the version names, so
a simple sort won't work. As a workaround we assume that any December date
comes from 2014 so it sorts before Jan, Feb, etc. Hopefully we'll add a date
to the version names before we hit Dec 2015 so we'll be able to take this
hack out.

This fixes a problem with the route/stack/etc. information not being displayed
on errors because we're not actually looking at the latest version.

Test Plan: Crossed fingers.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: jace

Differential Revision: https://phabricator.khanacademy.org/D15938