If we match an old expired error, update the error's key in the error def.
Summary:
If the error information has expired from Redis, then it's possible that the
error key will match on a new error. Since the new error will have a new def
created for it, that will cause the error key in the def to not match the key
that it's stored in. This updates the key in the def to match the error key.

Test Plan: Crossed fingers.

Reviewers: csilvers, jlfwong

Reviewed By: jlfwong

Subscribers: jace

Differential Revision: https://phabricator.khanacademy.org/D16126