Fix erroneous error categorization for prefix identifier matches.
Summary:
The intent of the various identifiers was to match errors by their
entire message or by parts of the message. However I messed up the
prefix identifier which was meant to be the first 3 words but instead
only included the error code/status and the first word, so any two
messages that began with the same word would be grouped. Oops!

This fixes things so that the first three words are used as the prefix
identifier.

Test Plan:
In Python shell, import models.py and call the following:

models._parse_message("A B C D E F G", 200, "3")

The returned error definition will have "200 3 A B C" in the `id1` field.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: jace

Differential Revision: https://phabricator.khanacademy.org/D16827