Handle occasional error where BigQuery returns an incomplete reponse.
Summary:
After placing some debug code on error-monitor-db in production I was
able to see what the return value is when there is no results data in
BigQuery's response. It seems like the job is marked incomplete even
though using the `bq` tool the job does show that it did complete (in 12
seconds).

For now, the workaround is just to retry since this happens infrequently
(a few times a day).

Test Plan: Crossed fingers.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: jace

Differential Revision: https://phabricator.khanacademy.org/D16901