Use the last line of the error message as the title when the first is empty.
Summary:
There are a bunch of errors that all seem to have an empty first line in the
error message string where the error description usually goes. It seems like
all exception instances that don't have a message (for instance,
DeadlineExceeded errors) fall into this category. The name of the exception
is found at the end, after the stack trace, so we use that as the title
instead.

Test Plan:
- Added some logging in the new `if` clause.
- Run bigquery_import.py
- See that the new error titles are more useful than empty string

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: jace

Differential Revision: https://phabricator.khanacademy.org/D17025