Fix overall count to no longer be ever-increasing
Summary:
When visiting /devadmin/errors, the table shows you a list of all the errors
that have happened in the past week along with the overall count of how
many times each error happened that week. The problem is that the overall
count is actually the count over all time, not over the past week. This is
confusing and also deceptive because newly occuring frequent errors may be
overshadowed by old errors that occur a lot.

The root cause of this issue is that the `<error_key>:versions` entry is a
sorted set of the versions for which an error has occurred. The key itself
has its expiration extended each time an error occurs, but what we actually
want is //each key// within the sorted set to expire after a certain amount
of time. The `ver:<version_key>:error:<error_key>:hours_seen` redis key
stores nearly identical information, but is expired out correctly. The end
result is that when we call `/api/internal/dev/errors/recent` the
`byHourAndVersion` member has everything we need and not more, but the
`versions` member has a lot of extra.

So, I added some code to manually expire out the keys in `<error_key>:versions`
whenever we call `/api/internal/dev/errors/recent`. Now, the overall counts are
what you expect them to be, and we're no longer growing without bound in redis!

Test Plan: Run the new code on internal-services, modify my local dev appserver to connect to it, and hit /devadmin/errors. Notice that the new counts are now correct.

Reviewers: tom

Reviewed By: tom

Differential Revision: https://phabricator.khanacademy.org/D17308