Fix first_seen to be a rolling list of log_hour
Summary:
Previously, we would keep track of the first time an error message was seen and keep that time around for one week. While that week-long expiration was passing new timestamps would come in but the `first_seen` would never be overwritten because all of the new timestamps were larger than the one already stored. When the `first_seen` key finally expires the next largest timestamp becomes the new `first_seen`, bypassing all of the timestamps that came in during the interim. The result was that the `first_seen` would jump from "8 days ago" to "3 hours ago" even tho there had been a steady stream of errors in the interim.

This changeset changes the `first_seen` key to a sorted set and manually ages out keys as they come in.

I did not put any effort in to migrating the current `first_seen` keys since they are pretty broken anyway. It will take ~1 week for the new `first_seen` keys to be correct and then I can also remove the `TODO` code.

Test Plan: Mainly playing around in python REPL, but I will probably do some more testing after deploying.

Reviewers: tom

Reviewed By: tom

Subscribers: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D17521