Make sure we always have something we can click on in the msg.
There's a mapreduce error that yields an empty firstline to the
error-monitor-db.  It has actual content in it, so I think there's
actually a bug in the error-monitor-db somewhere, but this will let us
click on the mapreduce error while we figure out the other bug.

Auditors: tom

Test plan:
Fingers crossed