Slackify report_errors.py
Summary:
This wasn't too complicated, but it got a little tricky to try to format the
hipchat, slack, and slack fallback messages prettily, without either
copy-pasting too much or overengineering the thing to death.  But it should do
the job, and when we remove hipchat we can just remove a bunch of stuff.
Unfortunately, the code length increased significantly, but some of that I
think improves code readability, some improves formatting in slack, and a lot
of it will go away when hipchat does.

Test Plan:
With a symlinked secrets.py:
`./report_errors.py --host internal-services.khanacademy.org --hipchat "HipChat Tests" --slack "#bot-testing"`

Reviewers: benjaminpollack, mroth

Reviewed By: benjaminpollack, mroth

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D21199