Don't alert about OOMs and request timeouts on a deploy
Summary:
We'll still alert if there are a whole lot of them, but a few OOMs or "request
was aborted after waiting too long"s on a deploy are unfortunately common and
probably not the deployer's fault.  This should make AlertiGator a good bit
less whiny in a fairly simple and safe way.

Test Plan: fingers crossed.

Reviewers: sean, csilvers

Reviewed By: sean, csilvers

Differential Revision: https://phabricator.khanacademy.org/D22895