Fix tests and requirements.
Summary:
The tests had bitrotted quite a bit.  For instance, we no longer
report an error that has occurred only once, but several of our tests
assumed that we did.  And some tests tested a return value that had
disappeared a long time ago!  I brought it back.

The biggest sadness in the test-fixing is that we can no longer test
the /error/<key> route because we clear the relevant <key> too
quickly, in verify_versions.  (The key is saved when we do
record_occurrence_from_logs, which sets up hours_seen, but not when we
do record_occurrence_during_monitoring, which is what our test does.)
I don't know if this is intended behavior or not, but it only affects
the tests, not real operation, so I just removed those tests.

I got rid of a dependency on scipy, which is a pain to install.  I now
document our dependence on numpy.

Test Plan: python test_server.py

Reviewers: tom

Reviewed By: tom

Differential Revision: https://phabricator.khanacademy.org/D25759