Store the most recent error with a given key in redis, not the first.
Summary:
This gives us a better idea of what kinds of errors are happening
*now* with a given key.  This is important for keys that are pretty
generic, such as `error importing module ...` or `object None has no
property ...`.

Update only the error message and stats, not the id's.

We need to keep all the id's the same to keep the error-key the same.

I *think* this means that an error-def can have now id's that are
unrelated to the actual error-message (error_def['title']) whereas
before that was not possible.  That is, we could have
```
error_def = { 'title': 'Found an unexpected key "george"',
              'status': 500,
              'level': 4,
              'id0': '200 4 Found an unexpected value "sam",
              'id1': '200 4 Found an unexpected',
              'id2': '200 4 unexpected value "sam"',
              'id3': None,
            }
```
where the idX fields do not match the title at all.

However, before it *was* possible for the count for an error to
include messages that were unrelated to some of the id's, so it's
pretty much the same.

Test Plan:
./test_server.py

./test_server.py

Reviewers: tom

Reviewed By: tom

Differential Revision: https://phabricator.khanacademy.org/D25793