Use a Poisson model instead of a Gaussian for deploy-time error monitoring
Summary:
For low frequency errors, deploy-time error monitoring can be a bit spammy.  My
hypothesis is that this is because when we have a small number of errors, the
error rate is not well modeled by a Gaussian distribution, even in the ideal
case.  Error counts within a time interval (in the ideal case where waiting
time between events is exponential) should be Poisson distributed.  This diff
changes the error monintor to use a Poisson distribution to detect elevated
error rates.  For large error counts, this just converges to the Gaussian
distribution, so the behavior should be the same for the cases where the old
monitor was working well.

My plan is to put a note in the 1s and 0s deploys channel when I deploy this
for people to ping me if the error monitoring seems more spammy or otherwise
wrong (as well as keep an eye on deploys) to make sure this is working as
expected.

Test Plan:
- `pip install nose`
- `nosetests`

Reviewers: tom, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D26262