Updated models to detect errors from distinct IPs
Summary:
Updated record_occurrence_during_monitoring to store a new value in Redis
recording the number of times an error has been seen on distinct IP addresses
as well as an intermediate value for calculating this that stores the number
of times an error has been seen at a particular IP address. Updated get_monitoring_errors to return the unique by ip error count instead of the total error count.

Test Plan:
For unit tests, test_server was updated to test 1) that a large number of errors from a single ip address does not create an 'elevated error' and 2) that a large number of errors from distinct ip addresses does create an 'elevated error'.

Most testing for this plan involved analysis of whether this change would be beneficial, and the following was done:

To estimate the reduction of 'elevated errors', a script was used to analyze bigquery logs, calculating the number of 'elevated errors' counting by distinct ips and counting by total errors. When not looking only at deploy times, the 'elevated error' rate fell by 10-20% depending on the dates of the logs. When looking specifically at the 10 minute windows after a deploy, set default, analyzing 6 days of bigquery logs, the 'elevated error' rate fell by about 33%.

Finally, I tested whether 'false negatives' were created where a deploy was aborted because of errors after set default, but counting by distinct ips would have caused no errors to be reported. The number of 'set default' followed by 'abort' commands is very low; there have been 5 from 05/14/2016 to 06/15/2016. Each of these aborts was analyzed manually. In 3 cases, the number of distinct IPs causing the significant (manually labaled as such) error, was > 15, and thus, even counting by distinct ips, it would have constituted an 'elevated error'. In 1 case, the errors were in sentry logs, which are not affected by this change, and in 1 case, the specific errors could not be found on bigquery. Additionally, in multiple of these deploys, end to end tests failed as well, indicating that the deploy might have been aborted even if there had not been reported 'elevated errors'.

Based on the potential reduction of 'false positive' 'elevated errors' and the lack of introduced 'false negatives', this change seems to be beneficial for deploy error monitoring. One risk is that 'false negatives' might occur at some point. While the small number of observed set defaults followed by aborts means that I cannot confidently state that false negatives will not occur, the benefits of a large reduction in false positives seem to outweigh this risk.

Reviewers: tom

Reviewed By: tom

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D28079