Blacklist memcache set errors from a known problematic key
Summary:
See the comment; I've got a ticket in the infra backlog to solve this properly,
but it's probably nontrivial, so in the meantime it's not useful to complain on
every deploy.  I had to add regex support to the blacklist since we put the
namespace before the key in the error log.

Test Plan: cross fingers

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D30861