Use the log completion API to determine whether to load logs from bigquery
Summary:
In a number of previous diffs, I've changed our log export process so
that now logs are streamed directly into their hourly and daily tables
from dataflow without an extra step in the middle needed to query
recently streamed logs and put them in the correct table.  However,
this means that logs will start appearing in the destination tables
immediately at the start of each hour, so we can't use the presence or
absence of those tables to determine whether to load logs into the
error monitor.

D37407 added an API for querying whether a given hour of logs is
complete.  This revision switched us over to query that before
attempting to load the logs to the error monitor.

I also removed the code for loading an entire day of logs at a time
from, the main codepath.  I don't believe this was ever doing the right
thing any more in prod.  Because it was running after we already tried
the individual hours, and checking to see if the first hour was absent,
this would almost certainly only run if we were missing the first hour
of logs in a day, but not subsequent hours, which is never the right
thing to do.  The only time this would be useful is if we're manually
loading historical logs form after our hourly logs have expired.  I
changed this to be accessible via a command-line flag instead.

Test Plan:
- `python bigquery_import.py`
- verify it errors when 404ing from the API since it hasn't been
  deployed yet and does not try to run the corresponding queries.
- `python bigquery_import.py --use-daily-tables`
- verify it tries to load logs in from today's bigquery table
- will test things end-to-end once the log completion API is deployed

Reviewers: csilvers, tom

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D37421