Add Additional Errors to Deploy Blacklist
Summary:
Adding two additional errors to the blacklist for deploy error monitoring.

First is `Process terminated because the request deadline was exceeded` which
was posted to slack 100 times between April 1st and May 31st. It comes most
often from `/_ah/queue/deferred` and is likely from traffic getting routed to
new machines. It has a threshold of 100, which would have reduced the number
of slack posts to about half, but with the newest change slowing the rate of
transferring traffic, that is likely enough to cover it. It may need tweaking
going forward based on how the slower deploy goes.

The second is `DeadlineExceededError: The overall deadline for responding`
which is a similar error as above that also comes a lot from the same route.
This posted 90 times between the same dates, and the threshold of 50 would
have prevented all of those slack posts.

Test Plan:
Monitor deploys and errors to make sure that actionable errors are still making
it through

Reviewers: benkraft, csilvers, jacqueline

Reviewed By: benkraft, csilvers

Subscribers: amos

Differential Revision: https://phabricator.khanacademy.org/D45398