Raising Blacklist Threshold for "Request was aborted"
Summary:
I'm bumping up the "Request was aborted after waiting too long" blacklist threshold.

Between 2018/07/01 and 2018/09/06, this error has alerted to Slack 174 times. The trigger values for this were between 101 (1 over the threshold) and 4017.

I'm bumping this value up to 800, which should cover 70+% of the current cases being reported.

However, that's a very high threshold to have for a very visible user facing 500 error. We should work on bringing that number down into a more comfortable range

Test Plan: Check the slack channel and make sure the threshold adjusts properly and posts about it go down.

Reviewers: amos, benkraft, csilvers

Reviewed By: amos, benkraft, csilvers

Differential Revision: https://phabricator.khanacademy.org/D48325